in-line the memory fetching
