+++ aiuser/context/assembler.py
@@ -15,7 +15,7 @@ from aiuser.context.entry import (
     SYSTEM_NAME_SUMMARY,
     MessageEntry,
 )
-from aiuser.context.memory.retriever import MemoryRetriever
+from aiuser.context.memory import fetch_relevant_memory
 from aiuser.utils.utilities import format_variables, mention_to_text
 
 if TYPE_CHECKING:
@@ -116,8 +116,9 @@ class ConversationAssembler:
     async def _fetch_relevant_memory(self) -> Optional[str]:
         if not await self.services.config.guild(self.guild).query_memories():
             return None
-        retriever = MemoryRetriever(self.ctx, db=self.services.memories)
-        return await retriever.fetch_relevant(mention_to_text(self.init_message))
+        return await fetch_relevant_memory(
+            self.ctx, self.services.memories, mention_to_text(self.init_message)
+        )
 
     # --- message conversion / filtering ---
